ref: Move getExactDuration() into utils/duration/
