+++ static/app/components/duration.tsx
@@ -1,5 +1,5 @@
 import getDuration from 'sentry/utils/duration/getDuration';
-import {getExactDuration} from 'sentry/utils/formatters';
+import {getExactDuration} from 'sentry/utils/duration/getExactDuration';
 
 interface Props extends React.HTMLAttributes<HTMLSpanElement> {
   seconds: number;
